Extract CreateNFTModalContent to reusable form, and actually use it
