+++ src/components/CreateNFTForm/styles.scss
